Voting is started. (Not in the UI yet.)

+++ app/controllers/submissions_controller.rb
@@ -22,8 +22,32 @@ class SubmissionsController < ApplicationController
 
   def destroy
     submission = Submission.find(params[:id])
-    submission.destroy if submission.can_delete?(current_user)
+    if signed_in?
+      submission.destroy if submission.can_delete?(current_user)
+    end
     redirect_back_or_default
+  end
+
+  def vote
+    submission = Submission.find(params[:id])
+    if signed_in?
+      value = params[:value].to_i
+      Submission.transaction do
+        submission.votes.build do |vote|
+          vote.submission = submission
+          vote.user = current_user
+          vote.value = value
+        end
+        logger.info "HI"
+        logger.info submission.inspect
+        logger.info submission.votes.inspect
+        logger.info submission.votes.first.valid?
+        logger.info submission.votes.first.errors.inspect
+        submission.vote_total += value
+        submission.save!
+      end
+    end
+    redirect_back_or_default submission_path(submission)
   end
 
   private
